implemented generic dfdx az model
