+++ az-discrete-opt/Cargo.toml
@@ -6,13 +6,13 @@ edition = "2021"
 [dependencies]
 rand = "0.8.5"
 rand_distr = "0.4.3"
+dfdx = { version = "0.13.0", default-features = false, features = ["std", "fast-alloc", "cuda", "cudnn"]}
 
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 # [dependencies]
 
 # [dev-dependencies]
-# dfdx = { version = "0.13.0", default-features = false, features = ["std", "fast-alloc", "cuda", "cudnn"]}
 # priority-queue = "1.3.2"
 # rand = "0.8.5"
 # rayon = "1.7.0"
