+++ az-discrete-opt/src/lib.rs
@@ -1,3 +1,4 @@
+#![feature(slice_flatten)]
 #![feature(maybe_uninit_uninit_array)]
 #![feature(maybe_uninit_array_assume_init)]
